Refactor checkout flow and remove CheckoutContext, update service details and layout components
