+++ mobile/src/app/(checkout)/_layout.tsx
@@ -1,31 +1,6 @@
 import { Stack } from 'expo-router'
-import { CheckoutProvider } from '../../context/CheckoutContext'
-import { View, Text, StyleSheet, TouchableOpacity } from 'react-native'
-import { Ionicons } from '@expo/vector-icons'
-import { Theme } from '@/src/types/theme'
+import { View } from 'react-native'
 import { useCustomTheme } from '@/src/context/ThemeContext'
-
-const HeaderAddress = () => {
-	return (
-		<View
-			style={{
-				backgroundColor: 'red',
-				flexDirection: 'row',
-				alignItems: 'center',
-				justifyContent: 'space-between',
-				paddingHorizontal: 16,
-				paddingVertical: 12,
-				width: '100%',
-				flex: 1,
-			}}
-		>
-			<Ionicons name="chevron-back" size={24} color="white" />
-			<Text style={{ color: 'white', fontSize: 16, fontWeight: 'bold' }}>
-				Endereço
-			</Text>
-		</View>
-	)
-}
 
 export default function CheckoutLayout() {
 	const { theme } = useCustomTheme()
@@ -55,12 +30,3 @@ export default function CheckoutLayout() {
 		</Stack>
 	)
 }
-
-const useStyles = (theme: Theme) =>
-	StyleSheet.create({
-		text: {
-			color: theme.colors.text,
-			fontSize: 16,
-			fontWeight: 'bold',
-		},
-	})
